Add Enemy HP Display and damage.
Add collision damage for ammo base prefabs when hitting enemy game
objects.

+++ src/Weapons/Assets/Scripts/Enemies/BaseEnemy.cs
@@ -7,7 +7,7 @@ public class BaseEnemy : MonoBehaviour
     public float BaseSpeed = 1.0f;
     public float BaseDamage = 1.0f;
     
-    public float BaseHealth = 1.0f;
+    public int BaseHealth = 1;
 
     #endregion
